+++ src/main/java/dinkplugin/message/DiscordMessageHandler.java
@@ -9,6 +9,7 @@ import dinkplugin.message.templating.Template;
 import dinkplugin.notifiers.data.NotificationData;
 import dinkplugin.util.DiscordProfile;
 import dinkplugin.util.Utils;
+import dinkplugin.util.WorldUtils;
 import lombok.NonNull;
 import lombok.extern.slf4j.Slf4j;
 import net.runelite.api.Client;
@@ -240,6 +241,16 @@ public class DiscordMessageHandler {
         }
 
         NotificationBody.NotificationBodyBuilder<?> builder = mBody.toBuilder();
+
+        if (config.includeLocation()) {
+            if (mBody.getWorld() == null) {
+                builder.world(client.getWorld());
+            }
+
+            if (mBody.getRegionId() == null) {
+                builder.regionId(WorldUtils.getLocation(client).getRegionID());
+            }
+        }
 
         if (config.sendDiscordUser()) {
             builder.discordUser(DiscordProfile.of(discordService.getCurrentUser()));